build tag for probabilistic tests
